Cleanup and minor changes to resolve compiler warnings.
* src/filelist_ctrl.cpp
  (#include): removed unused xpms

* src/workbench.cpp
  (Workbench::Data::RemoveProject,
   Workbench::GetContext): use "int" instead of "size_t"
   since wxArrayString::Index returns one.




git-svn-id: http://rapidsvn.tigris.org/svn/rapidsvn/trunk@6222 65390229-12b7-0310-b90b-f21a5aa7ec8e

+++ src/workbench.cpp
@@ -77,7 +77,7 @@ public:
   bool
   RemoveProject (const char * path)
   {
-    size_t index = projects.Index (path);
+    int index = projects.Index (path);
 
     if (index == wxNOT_FOUND)
       return false;
@@ -157,7 +157,7 @@ Workbench::GetContext (const size_t index)
 svn::Context *
 Workbench::GetContext (const char * path)
 {
-  size_t index = m->projects.Index (path);
+  int index = m->projects.Index (path);
 
   if (index == wxNOT_FOUND)
     return 0;
